+++ src/Mirakurun/Channel.ts
@@ -107,8 +107,13 @@ export default class Channel {
                 return;
             }
 
-            if (channel.satelite && typeof channel.satelite !== "string") {
-                log.error("invalid type of property `satelite` in channel#%d (%s) configuration", i, channel.name);
+            if (channel.satelite && !channel.satellite) {
+                log.warn("renaming deprecated property name `satelite` to `satellite` in channel#%d (%s) configuration", i, channel.name);
+                (<any> channel).satellite = channel.satelite;
+            }
+
+            if (channel.satellite && typeof channel.satellite !== "string") {
+                log.error("invalid type of property `satellite` in channel#%d (%s) configuration", i, channel.name);
                 return;
             }