+++ src/Mirakurun/api/channels/{type}/{channel}/stream.ts
@@ -59,10 +59,13 @@ export const get: Operation = (req, res) => {
     let requestAborted = false;
     req.once("close", () => requestAborted = true);
 
+    const userId = (req.ip || "unix") + ":" + (req.connection.remotePort || Date.now());
+
     channel.getStream({
-        id: (req.ip || "unix") + ":" + (req.connection.remotePort || Date.now()),
+        id: userId,
         priority: parseInt(req.get("X-Mirakurun-Priority"), 10) || 0,
         agent: req.get("User-Agent"),
+        url: req.url,
         disableDecoder: (<number> <any> req.query.decode === 0)
     })
         .then(stream => {
@@ -74,6 +77,7 @@ export const get: Operation = (req, res) => {
             req.once("close", () => stream.emit("close"));
 
             res.setHeader("Content-Type", "video/MP2T");
+            res.setHeader("X-Mirakurun-Tuner-User-ID", userId);
             res.status(200);
             stream.pipe(res);
         })
@@ -86,7 +90,12 @@ get.apiDoc = {
     produces: ["video/MP2T"],
     responses: {
         200: {
-            description: "OK"
+            description: "OK",
+            headers: {
+                "X-Mirakurun-Tuner-User-ID": {
+                    type: "string"
+                }
+            }
         },
         404: {
             description: "Not Found"